requirement.txt
PyYAML
fuzzyfinder
ply
requests_toolbelt
bs4
jedi-language-server
tree_sitter
